rpc: migrate JSONRPCRequest functionality into request.cpp
