+++ src/rpc/rawtransaction_util.cpp
@@ -11,7 +11,7 @@
 #include <keystore.h>
 #include <policy/policy.h>
 #include <primitives/transaction.h>
-#include <rpc/protocol.h>
+#include <rpc/request.h>
 #include <rpc/util.h>
 #include <tinyformat.h>
 #include <univalue.h>
